Check login truoc khi vao cac page

+++ controllers/jobController.js
@@ -51,136 +51,93 @@ const upload = multer({ storage });
 
 // @Routes
 router.get('/manage', (req, res) => {
-	if (req.user.account_type == 1)
-	{
-		Job.find({org_id: req.user.account_id}, (err, jobs) => {
-			//console.log(jobs);
-			Org.findOne({_id: req.user.account_id}, (err, org) => {
-				res.render('jobs/orgs/manage', {
-					title: 'CodeDao | Manage Jobs',
-					account_type: req.user.account_type,
-					account_id: req.user.account_id,
-					currentAcc: org,
-					jobs: jobs
+	if(!req.isAuthenticated()) {
+		res.redirect('/login');
+	} else {
+		if (req.user.account_type == 1)
+		{
+			Job.find({org_id: req.user.account_id}, (err, jobs) => {
+				//console.log(jobs);
+				Org.findOne({_id: req.user.account_id}, (err, org) => {
+					res.render('jobs/orgs/manage', {
+						title: 'CodeDao | Manage Jobs',
+						account_type: req.user.account_type,
+						account_id: req.user.account_id,
+						currentAcc: org,
+						jobs: jobs
+					});
 				});
 			});
-		});
-	}
-	else 
-	{
-		res.redirect("/");
+		}
+		else
+		{
+			res.redirect("/");
+		}
 	}
 });
 
 // job dashboard - view all jobs except yours
 router.get('/', (req, res) => {
-	if (req.user.account_type == 1) 
-	{
-		Job.find({org_id : {$ne: req.user.account_id}}, (err, jobs) => {
-			Org.findOne({_id: req.user.account_id}, (err, org) => {
-				res.render('jobs/dashboard', {
-					title: 'CodeDao | Jobs',
-					account_type: req.user.account_type,
-					account_id: req.user.account_id,
-					currentAcc: org,
-					criteriaList: org.hashtags,
-					jobs: jobs
-				}); 
+	if(!req.isAuthenticated()) {
+		res.redirect('/login');
+	} else {
+		if (req.user.account_type == 1)
+		{
+			Job.find({org_id : {$ne: req.user.account_id}}, (err, jobs) => {
+				Org.findOne({_id: req.user.account_id}, (err, org) => {
+					res.render('jobs/dashboard', {
+						title: 'CodeDao | Jobs',
+						account_type: req.user.account_type,
+						account_id: req.user.account_id,
+						currentAcc: org,
+						criteriaList: org.hashtags,
+						jobs: jobs
+					});
+				});
 			});
-		});
-	}	
-	else
-	{
-		Job.find({}, (err, jobs) => {
-			User.findOne({_id: req.user.account_id}, (err, user) => {
-				res.render('jobs/dashboard', {
-					title: 'CodeDao | Jobs',
-					account_type: req.user.account_type,
-					account_id: req.user.account_id,
-					currentAcc: user,
-					criteriaList: user.interests.concat(user.skills),
-					jobs: jobs
-				}); 
+		}
+		else
+		{
+			Job.find({}, (err, jobs) => {
+				User.findOne({_id: req.user.account_id}, (err, user) => {
+					res.render('jobs/dashboard', {
+						title: 'CodeDao | Jobs',
+						account_type: req.user.account_type,
+						account_id: req.user.account_id,
+						currentAcc: user,
+						criteriaList: user.interests.concat(user.skills),
+						jobs: jobs
+					});
+				});
 			});
-		});
+		}
 	}
 });
 
 // create new job
 router.get('/create', (req, res) => {
-	if (req.user.account_type != 1)
-		res.redirect('/');
-	else
-		Org.findOne({_id: req.user.account_id}, (err, org) => {
-			res.render('jobs/orgs/create', {
-				title: 'CodeDao | Create A New Job',
-				account_type: req.user.account_type,
-				account_id: req.user.account_id,
-				currentAcc: org
+	if(!req.isAuthenticated()) {
+		res.redirect('/login');
+	} else {
+		if (req.user.account_type != 1)
+			res.redirect('/');
+		else
+			Org.findOne({_id: req.user.account_id}, (err, org) => {
+				res.render('jobs/orgs/create', {
+					title: 'CodeDao | Create A New Job',
+					account_type: req.user.account_type,
+					account_id: req.user.account_id,
+					currentAcc: org
+				});
 			});
-		});
+	}
 });
 
 router.post('/create', (req, res) => {
 	//console.log(req.body);
-
-	let data = req.body;
-	let name = data.name;
-	let org_id = req.user.account_id;
-	let org_name = data.org_name;
-	let desc = data.desc;
-	let hashtags = data.hashtags;
-	let app_form = data.app_form;
-	let app_deadline = data.app_deadline;
-	let facebook = data.facebook;
-	let website = data.website;
-	let jobImage = data.jobImage;
-
-	var newJob = new Job({
-		_id: new mongoose.Types.ObjectId(),
-		created_at: new Date(),
-		updated_at: new Date(),
-		name: name,
-		org_id: org_id,
-		org_name: org_name,
-		desc: desc,
-		hashtags: hashtags,
-		app_form: app_form,
-		app_deadline: app_deadline,
-		facebook: facebook,
-		website: website,
-		jobImage: jobImage
-	});
-
-	newJob.save((err, jobs) => {
-		if(err) {
-			console.log(err);
-			return;
-		  }
-		  console.log(jobs);
-	});
-
-	res.redirect('/jobs/manage');
-});
-
-// edit an existing job
-router.get('/manage/edit/:ID', (req, res) => {
-	Job.findOne({_id: req.params.ID}, (err, job) => {
-		Org.findOne({_id: req.user.account_id}, (err, org) => {
-			res.render('jobs/orgs/edit', {
-				title: 'CodeDao | Edit your Job',
-				account_type: req.user.account_type,
-				account_id: req.user.account_id,
-				currentAcc: org,
-				job: job
-			});
-		});
-	});
-});
-
-router.post('/edit', (req, res) => {
-		//console.log(req.body);
-
+	if(!req.isAuthenticated()) {
+		res.redirect('/login');
+	} else {
 		let data = req.body;
 		let name = data.name;
 		let org_id = req.user.account_id;
@@ -192,9 +149,9 @@ router.post('/edit', (req, res) => {
 		let facebook = data.facebook;
 		let website = data.website;
 		let jobImage = data.jobImage;
-	
+
 		var newJob = new Job({
-			_id: req.jobId,
+			_id: new mongoose.Types.ObjectId(),
 			created_at: new Date(),
 			updated_at: new Date(),
 			name: name,
@@ -208,7 +165,7 @@ router.post('/edit', (req, res) => {
 			website: website,
 			jobImage: jobImage
 		});
-	
+
 		newJob.save((err, jobs) => {
 			if(err) {
 				console.log(err);
@@ -216,17 +173,86 @@ router.post('/edit', (req, res) => {
 			  }
 			  console.log(jobs);
 		});
-	
-		res.redirect('/');
+
+		res.redirect('/jobs/manage');
+	}
+});
+
+// edit an existing job
+router.get('/manage/edit/:ID', (req, res) => {
+	if(!req.isAuthenticated()) {
+		res.redirect('/login');
+	} else {
+		Job.findOne({_id: req.params.ID}, (err, job) => {
+			Org.findOne({_id: req.user.account_id}, (err, org) => {
+				res.render('jobs/orgs/edit', {
+					title: 'CodeDao | Edit your Job',
+					account_type: req.user.account_type,
+					account_id: req.user.account_id,
+					currentAcc: org,
+					job: job
+				});
+			});
+		});
+	}
+});
+
+router.post('/edit', (req, res) => {
+		//console.log(req.body);
+		if(!req.isAuthenticated()) {
+			res.redirect('/login');
+		} else {
+			let data = req.body;
+			let name = data.name;
+			let org_id = req.user.account_id;
+			let org_name = data.org_name;
+			let desc = data.desc;
+			let hashtags = data.hashtags;
+			let app_form = data.app_form;
+			let app_deadline = data.app_deadline;
+			let facebook = data.facebook;
+			let website = data.website;
+			let jobImage = data.jobImage;
+
+			var newJob = new Job({
+				_id: req.jobId,
+				created_at: new Date(),
+				updated_at: new Date(),
+				name: name,
+				org_id: org_id,
+				org_name: org_name,
+				desc: desc,
+				hashtags: hashtags,
+				app_form: app_form,
+				app_deadline: app_deadline,
+				facebook: facebook,
+				website: website,
+				jobImage: jobImage
+			});
+
+			newJob.save((err, jobs) => {
+				if(err) {
+					console.log(err);
+					return;
+				  }
+				  console.log(jobs);
+			});
+
+			res.redirect('/');
+		}
 });
 
 //delete a job when requested
 router.post('/delete', (req, res) => {
-	console.log("DELETE!");
-	if (req.body.JobID != undefined) {
-		Job.findOneAndRemove({_id: req.body.JobID}, (err, job) => {});
+	if(!req.isAuthenticated()) {
+		res.redirect('/login');
+	} else {
+		console.log("DELETE!");
+		if (req.body.JobID != undefined) {
+			Job.findOneAndRemove({_id: req.body.JobID}, (err, job) => {});
+		}
+		res.redirect("/jobs/manage");
 	}
-	res.redirect("/jobs/manage");
 });
